+++ packages/docs/docs/main/change-log.md
@@ -12,6 +12,10 @@ This page lists the changes to the IsaacScript framework.
 ## February 14th, 2023 (Unreleased)
 
 - The "toggleDisplay" helper functions now accept an optional "force" argument.
+- Added the following helper functions:
+  - `sortObjectArrayByKey`
+- Renamed the following helper functions:
+  - `twoDimensionalSort` --> `sortTwoDimensional`
 - Added the following custom callbacks:
   - `POST_KEYBOARD_PRESSED`